api-pytest/rpm_auto-main/requirements.txt
allure-python-commons
pytest>=8.0.0
requests>=2.32.0
allure-pytest>=2.14.0
PyYAML>=6.0
